Refactor the code by interface.

+++ src/main/resources/public/scripts/deleteOrder.js
@@ -1,23 +1,25 @@
 $("form").submit(function(event){
     event.preventDefault();
 
-    var url = "../order";
     var orderId = $("#txtOrderId").val();
+    if(confirm("Do want to delete the order of " + orderId)){
+        var url = "../order";
 
-    $.ajax({
-        url : url + "?orderId=" + orderId,
-        success : function(data) {
-            if (data.message) {
-              alert("Delete Fail, " + data.message);
-            } else {
-              alert("Delete Successfully。");
-            }
-        },
+        $.ajax({
+            url : url + "?orderId=" + orderId,
+            success : function(data) {
+                if (data.message) {
+                  alert("Delete Fail, " + data.message);
+                } else {
+                  alert("Delete Successfully。");
+                }
+            },
 
-        error : function(data, textStatus, errorThrown) {
-            alert("Delete Fail, " + data.responseJSON.message);
-        },
-        type : "DELETE",
-    });
+            error : function(data, textStatus, errorThrown) {
+                alert("Delete Fail, " + data.responseJSON.message);
+            },
+            type : "DELETE",
+        });
+        };
     }
 );
